Restructure File System, Add README.md
To provide a more consitent workflow, the files have been put in file type
directories so as to have easy access and modifications of files as needed.
Previously, one would have to search through the entire DSS interface dir,
making it hard to find a specific file or manage links as the project scales.

The previous README only dealt with the database, so there was a need to create
another that addresses the project as a whole, providing valuable information
to people interested in the project. The README dealing with the database has
therefore been relocated to the database directory for the same reason.