Add index.php
This file, sitting at the base of the repository, has been added to enable the
redirection of the user to /interfaces/homepage.html, where all front-facing
UI/UX is. This is due to Web servers behaviour of searching for index.* at the
base of projects/ repositories and displaying everything as a tree when no
index.* file has been found.